+++ til_bigquery/__init__.py
@@ -1,3 +1,4 @@
+from .constants import BigQueryLocation
 from .exceptions import BigQueryFetchError, BigQueryInsertError
 from .model import BigQueryModel
 from .repository import BigQueryRepository