+++ til_bigquery/constants.py
@@ -1,12 +1,12 @@
 from enum import Enum
 
 
-class BigQueryLocation(Enum):
+class BigQueryLocation(str, Enum):
     EU = "EU"
     US = "US"
 
 
-class BigQueryMode(Enum):
+class BigQueryMode(str, Enum):
     NULLABLE = "NULLABLE"  # Optional
     REQUIRED = "REQUIRED"
     REPEATED = "REPEATED"  # List